refactor: separate component for the form
